+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Setup of CI with Github Actions
 - Setup of Dockerfile for local tests
 - Creation of CHANGELOG.md
+- Setup of CodeCov for CI
 
 ## [1.0.0] - 2021-04-15